+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Lexing of conditional directive expressions containing compiler directives, comments, or strings.
+
 ## [0.3.0] - 2024-05-29
 
 ### Removed